requirements.txt
altgraph==0.17.4
certifi==2024.8.30
charset-normalizer==3.4.0
diffusers==0.31.0
filelock==3.13.1
fsspec==2024.2.0
huggingface-hub==0.26.2
idna==3.10
importlib_metadata==8.5.0
Jinja2==3.1.3
macholib==1.16.3
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==2.1.3
packaging==24.2
pillow==11.0.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
sympy==1.13.1
tokenizers==0.20.3
torch==2.5.1
tqdm==4.67.0
transformers==4.46.2
typing_extensions==4.9.0
urllib3==2.2.3
zipp==3.20.2
